Rename parsers.py to commands.py
